Let 'onSuccess' and 'onFailure' accept a consumer which might throw an exception
If 'onSuccess' and 'onFailure' only accept a 'Consumer', it is extremely uncomfortable to use this 'Try' to call methods which might throw an exception. Using 'isSuccess' breaks fluency. I guess from a functional programming point of view calling an exception throwing method from a Try-monad could be considered to contradict the type's intention. But if using this monad in an existing non-FP code base, there might sometimes be no other choice. It would be a shame not to use Try in those situations.

With a 'TryConsumer' the methods transparently throw the same exception as the consumer.